api-service/requirements.txt
Django
djangorestframework
gunicorn
mysqlclient
requests
redis
